+++ homework/jeon-suwon/re-api/services/todoitem.service.js
@@ -1,0 +1,8 @@
+import * as todoitemRepository from "../repositories/todoitem.repository.js";
+
+export const putTodoitem = (id, userId) => {
+  const findTodoItem = todoitemRepository.findTodoItem(id);
+
+  const data = todoitemRepository.changeTodoItem(id, userId, findTodoItem);
+  return data;
+};
